Refactor: Split Medication Utils and Tests into Subdirectories
Further refactors the medication feature's utility functions and tests for better organization:

- Utility Files: Splits the `src/app/medication/utils.ts` file into:
    - `src/app/medication/utils/schedule-formatting.ts` (for `formatSchedule`)
    - `src/app/medication/utils/next-due-calculation.ts` (for `calculateNextDue`)
- Test Files: Splits the corresponding test file `src/app/medication/utils.test.ts` into:
    - `src/app/medication/utils/schedule-formatting.test.ts`
    - `src/app/medication/utils/next-due-calculation.test.ts`
- Imports: Updates all relevant import paths in components (`MedicationRegimenCard`, `RegimenAccordionContent`) to reflect the new locations of these utility modules.
- Verification: All formatting, linting, tests, and build processes passed successfully after these changes.
